requirements.txt
numpy>=1.21.0
Pillow>=8.3.1
matplotlib>=3.4.3
tqdm>=4.62.0
streamlit>=1.31.0
graphviz>=0.20.1 